#13 Implementation des différents

+++ App.Core/Services/CopyService.cs
@@ -14,11 +14,39 @@ namespace App.Core.Services
             {
                 if (File.Exists(model.SourcePath))
                 {
-                    // Copy file
-                    File.Copy(model.SourcePath, model.TargetPath, true);
-                    Console.WriteLine("File copied successfully.");
-                }
+                    // Check if it's a complete or differential save
+                    if (saveModel.Type == false || !File.Exists(saveModel.InPath))
+                    {
+                        // Complete save or source file not found for differential save
+                        File.Copy(model.SourcePath, model.TargetPath, true);
+                        Console.WriteLine("File copied successfully.");
+                    }
+                    else
+                    {
+                        // Differential save
+                        FileInfo sourceFileInfo = new FileInfo(model.SourcePath);
+                        FileInfo inFileInfo = new FileInfo(saveModel.InPath);
 
+                        // Check if source file has been modified since the last save
+                        if (sourceFileInfo.LastWriteTime > inFileInfo.LastWriteTime)
+                        {
+                            File.Copy(model.SourcePath, model.TargetPath, true);
+                            Console.WriteLine("File copied successfully.");
+                        }
+                        else
+                        {
+                            LoggerModel loggerModel = new LoggerModel();
+
+                            loggerModel.FileSource = model.SourcePath;
+                            loggerModel.FileTarget = model.TargetPath;
+                            loggerModel.FileSize = "";
+
+                            loggerModel.FileTransferTime = "00:00:00.0000000";
+                            loggerService.WriteLog(loggerModel, saveModel);
+                            Console.WriteLine("Source file has not been modified since the last save.");
+                        }
+                    }
+                }
                 else if (Directory.Exists(model.SourcePath))
                 {
                     // Source is a directory
@@ -26,17 +54,15 @@ namespace App.Core.Services
 
                     if (files.Length > 0)
                     {
-                        CopyDirectory(model.SourcePath,
-                                      model.TargetPath, saveModel);
+                        CopyDirectory(model.SourcePath, model.TargetPath, saveModel);
                         Console.WriteLine("Directory copied successfully.");
 
                         // Log an entry for each file in the directory
                         foreach (string filePath in files)
                         {
                             string destFilePath = Path.Combine(model.TargetPath, Path.GetFileName(filePath));
-                            
+
                         }
-                        
                     }
                     else
                     {
@@ -53,8 +79,7 @@ namespace App.Core.Services
                     loggerModel.FileSize = "";
 
                     loggerModel.FileTransferTime = "-1";
-                    loggerService.WriteLog(loggerModel, saveModel); ;
-                    
+                    loggerService.WriteLog(loggerModel, saveModel);
                 }
             }
             catch (Exception e)
@@ -62,6 +87,7 @@ namespace App.Core.Services
                 Console.WriteLine($"Error: {e.Message}");
             }
         }
+
